back-end/data-analysis/requirements.txt
scikit-learn
seaborn
scipy
fastapi
uvicorn
psycopg2
sqlalchemy
pytest
pandas
openpyxl
python-dotenv